requirements.txt
pyrofork
apscheduler
python-dotenv
tgcrypto
pymongo
motor
umongo
asyncio
colorama
aiohttp
pytz
IMDbPY
aiofiles
uvloop
pyshorteners
cloudscraper
psutil
requests
shortzy
ffmpeg-python
telegraph
marshmallow==3.26.1
httpx
